refactor: taskEnvironment 삭제 처리 및 관련 코드 수정
- viewModel 불필요한 주입코드 삭제
- taskEnvironment 코드 taskViewModel로 이동
